The two types of local modes are divided by the unit of source code.
