+++ aiaccel/master/create.py
@@ -3,6 +3,7 @@ from typing import Any
 from aiaccel.config import Config
 from aiaccel.master.abci_master import AbciMaster
 from aiaccel.master.local_master import LocalMaster
+from aiaccel.master.pylocal_master import PylocalMaster
 
 
 def create_master(config_path: str) -> Any:
@@ -12,10 +13,14 @@ def create_master(config_path: str) -> Any:
     config = Config(config_path)
     resource = config.resource_type.get()
 
-    if resource.lower() == "local" or resource.lower() == "python_local":
+    if resource.lower() == "local":
         return LocalMaster
+
+    elif resource.lower() == "python_local":
+        return PylocalMaster
 
     elif resource.lower() == "abci":
         return AbciMaster
+
     else:
         return None
