+++ aiaccel/master/pylocal_master.py
@@ -1,0 +1,8 @@
+from aiaccel.master.abstract_master import AbstractMaster
+
+
+class PylocalMaster(AbstractMaster):
+    """A master class running on a local computer.
+
+    """
+    pass
