+++ aiaccel/scheduler/local_scheduler.py
@@ -1,40 +1,18 @@
 import re
 import subprocess
 import threading
-import importlib
 
 from typing import Union
-from pathlib import Path
 
 from aiaccel.scheduler.abstract_scheduler import AbstractScheduler
-from aiaccel.util.aiaccel import Run
-from aiaccel.util.filesystem import create_yaml
 from aiaccel.util.process import ps2joblist
 from aiaccel.wrapper_tools import create_runner_command
-from aiaccel.util.time_tools import get_time_now
-from aiaccel import dict_result
 
 
 class LocalScheduler(AbstractScheduler):
     """A scheduler class running on a local computer.
 
     """
-
-    def __init__(self, options: dict) -> None:
-        super().__init__(options)
-
-        self.run = None
-        self.user_func = None
-
-        if self.is_python_local_mode():
-            self.user_func = self.get_callable_object(
-                self.config.python_file.get(),
-                self.config.function.get()
-            )
-            self.run = Run(self.config_path)
-
-    def is_python_local_mode(self):
-        return self.config.resource_type.get().lower() == 'python_local'
 
     def get_stats(self) -> None:
         """Get a current status and update.
@@ -81,22 +59,6 @@ class LocalScheduler(AbstractScheduler):
 
         return args[trial_id_index + index_offset]
 
-    def get_callable_object(self, file_path: Union[str, Path], attr_name: str) -> callable:
-        """ Loads the specified module from the specified python program.
-
-        Args:
-            file_path (str, pathlib.Path): A user program file path.(python file only)
-            attr_name (str): A name of objective function in user program.
-
-        Returns:
-            callable
-        """
-        spec = importlib.util.spec_from_file_location("user_module", file_path)
-        module = importlib.util.module_from_spec(spec)
-        spec.loader.exec_module(module)
-
-        return getattr(module, attr_name)
-
     def inner_loop_main_process(self) -> bool:
         """A main loop process. This process is repeated every main loop.
 
@@ -119,7 +81,7 @@ class LocalScheduler(AbstractScheduler):
         return True
 
     def execute(self, trial_id: int) -> None:
-        """ Execute in subprocess or directly from a loaded callable object.
+        """ Generates and executes commands to run user programs.
 
         Args:
             trial_id (int): Any trial od
@@ -129,25 +91,6 @@ class LocalScheduler(AbstractScheduler):
         """
         self.storage.trial.set_any_trial_state(trial_id=trial_id, state='running')
 
-        if self.is_python_local_mode():
-            self.execute_callable(trial_id)
-        else:
-            self.execute_in_subprocess(trial_id)
-
-        self.storage.trial.set_any_trial_state(trial_id=trial_id, state='finished')
-        self.create_result_file(trial_id)
-
-        return
-
-    def execute_in_subprocess(self, trial_id: int) -> None:
-        """ Generates and executes commands to run user programs.
-
-        Args:
-            trial_id (int): Any trial od
-
-        Returns:
-            None
-        """
         runner_command = create_runner_command(
             self.config.job_command.get(),
             self.storage.get_hp_dict(trial_id),
@@ -156,51 +99,7 @@ class LocalScheduler(AbstractScheduler):
         )
         subprocess.run(runner_command)
 
-        return
-
-    def execute_callable(self, trial_id: int) -> None:
-        """ Executes the loaded callable object.
-
-        Args:
-            trial_id (int): Any trial od
-
-        Returns:
-            None
-        """
-        start_time = get_time_now()
-        xs, y, err = self.run.execute(self.user_func, trial_id, y_data_type=None)
-        end_time = get_time_now()
-        self.run.report(trial_id, xs, y, err, start_time, end_time)
+        self.storage.trial.set_any_trial_state(trial_id=trial_id, state='finished')
+        self.create_result_file(trial_id)
 
         return
-
-    def create_result_file(self, trial_id: int) -> None:
-        """ Save the results in yaml format.
-
-        Args:
-            trial_id (int): Any trial od
-
-        Returns:
-            None
-        """
-
-        file_hp_count_fmt = f'%0{self.config.name_length.get()}d'
-        file_name = file_hp_count_fmt % trial_id + '.hp'
-
-        content = self.storage.get_hp_dict(trial_id)
-        result = self.storage.result.get_any_trial_objective(trial_id=trial_id)
-        error = self.storage.error.get_any_trial_error(trial_id=trial_id)
-
-        content['result'] = str(result)
-
-        if error is not None:
-            content['error'] = error
-
-        result_file_path = self.ws / dict_result / file_name
-        create_yaml(result_file_path, content)
-
-    def __getstate__(self):
-        obj = super().__getstate__()
-        del obj['run']
-        del obj['user_func']
-        return obj
